+++ tests/migrations/0013_authorization.py
@@ -20,7 +20,7 @@ class Migration(migrations.Migration):
             field=models.ForeignKey(
                 blank=True,
                 null=True,
-                on_delete=django.db.models.deletion.SET_NULL,
+                on_delete=django.db.models.deletion.RESTRICT,
                 related_name="%(app_label)s_%(class)s",
                 to=settings.OAUTH2_PROVIDER_AUTHORIZATION_MODEL,
             ),
@@ -31,7 +31,7 @@ class Migration(migrations.Migration):
             field=models.ForeignKey(
                 blank=True,
                 null=True,
-                on_delete=django.db.models.deletion.SET_NULL,
+                on_delete=django.db.models.deletion.RESTRICT,
                 related_name="%(app_label)s_%(class)s",
                 to=settings.OAUTH2_PROVIDER_AUTHORIZATION_MODEL,
             ),
@@ -42,7 +42,7 @@ class Migration(migrations.Migration):
             field=models.ForeignKey(
                 blank=True,
                 null=True,
-                on_delete=django.db.models.deletion.SET_NULL,
+                on_delete=django.db.models.deletion.RESTRICT,
                 related_name="%(app_label)s_%(class)s",
                 to=settings.OAUTH2_PROVIDER_AUTHORIZATION_MODEL,
             ),
@@ -53,7 +53,7 @@ class Migration(migrations.Migration):
             field=models.ForeignKey(
                 blank=True,
                 null=True,
-                on_delete=django.db.models.deletion.SET_NULL,
+                on_delete=django.db.models.deletion.RESTRICT,
                 related_name="%(app_label)s_%(class)s",
                 to=settings.OAUTH2_PROVIDER_AUTHORIZATION_MODEL,
             ),
@@ -75,7 +75,7 @@ class Migration(migrations.Migration):
             field=models.ForeignKey(
                 blank=True,
                 null=True,
-                on_delete=django.db.models.deletion.SET_NULL,
+                on_delete=django.db.models.deletion.CASCADE,
                 related_name="%(app_label)s_%(class)s",
                 to=settings.OAUTH2_PROVIDER_AUTHORIZATION_MODEL,
             ),
@@ -91,7 +91,7 @@ class Migration(migrations.Migration):
             field=models.ForeignKey(
                 blank=True,
                 null=True,
-                on_delete=django.db.models.deletion.SET_NULL,
+                on_delete=django.db.models.deletion.RESTRICT,
                 related_name="%(app_label)s_%(class)s",
                 to=settings.OAUTH2_PROVIDER_AUTHORIZATION_MODEL,
             ),